Added all security functionality and additional API functionality

+++ Yummgy/src/main/java/com/cognixia/jump/controller/UserController.java
@@ -5,19 +5,23 @@ import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.cognixia.jump.exception.ResourceNotFoundException;
 import com.cognixia.jump.model.Favorites;
+import com.cognixia.jump.model.Recipe;
 import com.cognixia.jump.model.User;
 import com.cognixia.jump.repository.UserRepository;
+import com.cognixia.jump.util.JwtUtil;
 
 import jakarta.validation.Valid;
 
@@ -28,6 +32,12 @@ public class UserController {
 	
 	@Autowired
 	UserRepository repo;
+	
+	@Autowired
+	PasswordEncoder encoder;
+	
+	@Autowired
+	JwtUtil jwtUtil;
 	
 	@CrossOrigin
 	@GetMapping("/users")
@@ -44,6 +54,8 @@ public class UserController {
 		
 		newUser.setUserId(null);
 		
+		newUser.setYumPassword( encoder.encode( newUser.getYumPassword() ) );
+		
 		User added = repo.save(newUser); 
 		
 		System.out.println("Added: " + added);
@@ -53,23 +65,56 @@ public class UserController {
 	
 	@CrossOrigin
 	@DeleteMapping("/delete/user/{id}")
-	public ResponseEntity<?> deleteUser(@PathVariable int id) throws ResourceNotFoundException {
+	public ResponseEntity<?> deleteUser(@RequestHeader (name="Authorization") String token, @PathVariable int id) throws ResourceNotFoundException {
 		
 		Optional<User> found = repo.findById(id);
 		
+		
 		if(found.isPresent()) {
 			
-			repo.deleteById(id);
-			
-			return ResponseEntity.status(200).body(found.get());	
+			if(found.get().equals(jwtUtil.getLoggedInUser(token)))
+			{
+				repo.deleteById(id);
+				
+				return ResponseEntity.status(200).body(found.get());	
+			}
+			else
+			{
+				return ResponseEntity.status(404).body("user not yours");
+			}
 		}
 		else {
 			throw new ResourceNotFoundException("User", id);
 		}
 			
 	}
+	
+	@CrossOrigin
+	@GetMapping("/users/recipes")
+	public ResponseEntity<?> getLoggedInUserRecipes(@RequestHeader (name="Authorization") String token) {
+		return ResponseEntity.status(200).body(jwtUtil.getLoggedInUser(token).getRecipes());
+	}
+	
+	@CrossOrigin
+	@GetMapping("/users/favorites")
+	public ResponseEntity<?> getLoggedInUserFavorites(@RequestHeader (name="Authorization") String token) {
+		return ResponseEntity.status(200).body(jwtUtil.getLoggedInUser(token).getFavorites());
+	}
+	
+	@CrossOrigin
+	@GetMapping("/users/{userId}/recipes")
+	public ResponseEntity<?> getUserRecipes(@PathVariable int id) throws ResourceNotFoundException {
+		Optional<User> userOptional = repo.findById(id);
 
+        if (userOptional.isPresent()) {
+            List<Recipe> recipes = userOptional.get().getRecipes();
+            return ResponseEntity.status(200).body(recipes);
+        } else {
+        	throw new ResourceNotFoundException("User", id);
+        }
+	}
     
+	@CrossOrigin
     @GetMapping("/users/{userId}/favorites")
     public ResponseEntity<?> getUserFavorites(@PathVariable Integer userId) throws ResourceNotFoundException {
         Optional<User> userOptional = repo.findById(userId);
